Lastpass OSX
Updating the lastpass cli to install for both linux and osx.

+++ bin/lastpass-cli.sh
@@ -5,27 +5,57 @@
 
 set -e
 
-sudo apt-get install -y \
-  asciidoc \
-  build-essential \
-  cmake \
-  libcurl4-openssl-dev \
-  libssl-dev \
-  libxml2 \
-  libxml2-dev \
-  openssl \
-  pinentry-curses \
-  pkg-config \
-  xclip \
-  xsltproc
+is_osx() {
+	local platform
+  platform=$(uname)
+	[ "$platform" == "Darwin" ]
+}
 
-git clone git@github.com:lastpass/lastpass-cli.git
+linux_install() {
+  sudo apt-get install -y \
+    asciidoc \
+    build-essential \
+    cmake \
+    libcurl4-openssl-dev \
+    libssl-dev \
+    libxml2 \
+    libxml2-dev \
+    openssl \
+    pinentry-curses \
+    pkg-config \
+    xclip \
+    xsltproc
 
-(
-cd lastpass-cli
-make
-sudo make install
-sudo make install-doc
-)
+  if [ -d 'lastpass-cli' ]; then
+    echo "Removing old install dir"
+    rm -rf lastpass-cli
+  fi
+  git clone git@github.com:lastpass/lastpass-cli.git
 
-rm -rf lastpass-cli
+  (
+  cd lastpass-cli
+  make
+  sudo make install
+  sudo make install-doc
+  )
+
+  rm -rf lastpass-cli
+}
+
+osx_install() {
+  brew update
+  brew install lastpass-cli --with-pinentry
+}
+
+main () {
+  if is_osx; then
+    echo 'Installing for OSX'
+    osx_install
+  else
+    echo 'Installing for Linux'
+    linux_install
+  fi;
+}
+
+main "$@"
+
